Extract device information from DefaultAndroidEventProcessor
