+++ sentry-android-core/src/main/java/io/sentry/android/core/ContextUtils.java
@@ -22,10 +22,39 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-final class ContextUtils {
+@ApiStatus.Internal
+public final class ContextUtils {
+
+  static class SideLoadedInfo {
+    private final boolean isSideLoaded;
+    private final @Nullable String installerStore;
+
+    public SideLoadedInfo(boolean isSideLoaded, @Nullable String installerStore) {
+      this.isSideLoaded = isSideLoaded;
+      this.installerStore = installerStore;
+    }
+
+    public boolean isSideLoaded() {
+      return isSideLoaded;
+    }
+
+    public @Nullable String getInstallerStore() {
+      return installerStore;
+    }
+
+    public @NotNull Map<String, String> asTags() {
+      final Map<String, String> data = new HashMap<>();
+      data.put("isSideLoaded", String.valueOf(isSideLoaded));
+      if (installerStore != null) {
+        data.put("installerStore", installerStore);
+      }
+      return data;
+    }
+  }
 
   private ContextUtils() {}
 
@@ -187,8 +216,8 @@ final class ContextUtils {
     return defaultVersion;
   }
 
-  @SuppressWarnings("deprecation")
-  static @Nullable Map<String, String> getSideLoadedInfo(
+  @SuppressWarnings({"deprecation"})
+  static @Nullable SideLoadedInfo retrieveSideLoadedInfo(
       final @NotNull Context context,
       final @NotNull ILogger logger,
       final @NotNull BuildInfoProvider buildInfoProvider) {
@@ -202,20 +231,10 @@ final class ContextUtils {
 
         // getInstallSourceInfo requires INSTALL_PACKAGES permission which is only given to system
         // apps.
+        // if it's installed via adb, system apps or untrusted sources
+        // could be amazon, google play etc - or null in case of sideload
         final String installerPackageName = packageManager.getInstallerPackageName(packageName);
-
-        final Map<String, String> sideLoadedInfo = new HashMap<>();
-
-        if (installerPackageName != null) {
-          sideLoadedInfo.put("isSideLoaded", "false");
-          // could be amazon, google play etc
-          sideLoadedInfo.put("installerStore", installerPackageName);
-        } else {
-          // if it's installed via adb, system apps or untrusted sources
-          sideLoadedInfo.put("isSideLoaded", "true");
-        }
-
-        return sideLoadedInfo;
+        return new SideLoadedInfo(installerPackageName == null, installerPackageName);
       }
     } catch (IllegalArgumentException e) {
       // it'll never be thrown as we are querying its own App's package.
