+++ sentry/api/sentry.api
@@ -655,11 +655,6 @@ public abstract interface class io/sentry/IntegrationName {
 	public fun getIntegrationName ()Ljava/lang/String;
 }
 
-public final class io/sentry/InternalSentrySdk {
-	public fun <init> ()V
-	public static fun serializeScope (Lio/sentry/Scope;)Ljava/util/Map;
-}
-
 public final class io/sentry/IpAddressUtils {
 	public static final field DEFAULT_IP_ADDRESS Ljava/lang/String;
 	public static fun isDefault (Ljava/lang/String;)Z
@@ -1179,7 +1174,10 @@ public final class io/sentry/Scope {
 	public fun clearAttachments ()V
 	public fun clearBreadcrumbs ()V
 	public fun clearTransaction ()V
+	public fun getBreadcrumbs ()Ljava/util/Queue;
 	public fun getContexts ()Lio/sentry/protocol/Contexts;
+	public fun getExtras ()Ljava/util/Map;
+	public fun getFingerprint ()Ljava/util/List;
 	public fun getLevel ()Lio/sentry/SentryLevel;
 	public fun getRequest ()Lio/sentry/protocol/Request;
 	public fun getSession ()Lio/sentry/Session;
